fix: make kato_id in SettlementInsertDB Long type

+++ src/main/kotlin/com/idf/kz/model/SettlementInsertDB.kt
@@ -1,10 +1,10 @@
 package com.idf.kz.model
 
 data class SettlementInsertDB(
-  val regionId: Int,
-  val addressDistrictId: Int,
-  val addressSettlementTypeId: Int,
+  val regionId: Long,
+  val addressDistrictId: Long,
+  val addressSettlementTypeId: Long,
   val name: String,
-  val katoId: Int,
+  val katoId: Long,
   val parentName: String,
   val resource: String)
